Revert "Feat/edu refactor 2"
