+++ src/App.tsx
@@ -1,5 +1,9 @@
 import { useEffect, useState } from "react";
-import { getLisitaCollection, db, lisitaDocumentCollection } from "./firebase";
+import {
+  getLisitaCollection,
+  db,
+  lisitaDocumentCollection,
+} from "./components/firebase";
 import PrintCollection from "./components/PrintCollection";
 import { MyForm } from "./components/MyForm";
 
@@ -7,7 +11,10 @@ function App() {
   const [documentsLisita, setDocumentsLisita] =
     useState<lisitaDocumentCollection[]>();
   useEffect(() => {
-    getLisitaCollection(db).then((data) => setDocumentsLisita(data));
+    getLisitaCollection(db).then((data) => {
+      //data.map((data)=>{console.log("data:", data)})
+      setDocumentsLisita(data);
+    });
   }, []);
   return (
     <>
